fix install 9

+++ install.sh
@@ -6,22 +6,34 @@ curl -O https://downloads.ioncube.com/loader_downloads/ioncube_loaders_lin_x86-6
 tar -xvzf ioncube_loaders_lin_x86-64.tar.gz
 rm ioncube_loaders_lin_x86-64.tar.gz
 cd ioncube
+
+
 php_version="$(php -r 'echo PHP_MAJOR_VERSION . "." . PHP_MINOR_VERSION;' 2>/dev/null)"
+
+if [ "$php_version" == "7.4" ]; then
+    php_ext_dir="$(php -r 'echo ini_get("extension_dir");' 2>/dev/null)"
+    cp ioncube_loader_lin_${php_version}.so /usr/lib/php/${php_ext_dir}
+    cd ..
+    rm -rf ioncube
+    cat > /etc/php/${php_version}/cli/conf.d/00-ioncube-loader.ini << EOF
+    zend_extension=ioncube_loader_lin_${php_version}.so
+EOF
+    else
     php_ext_dir="$(command php -i | grep extension_dir 2>'/dev/null' \
     | command head -n 1 \
     | command cut --characters=31-38)"
     php_version="$(command php --version 2>'/dev/null' \
     | command head -n 1 \
     | command cut --characters=5-7)"
-
-
-cp ioncube_loader_lin_${php_version}.so /usr/lib/php/${php_ext_dir}
-cd ..
-rm -rf ioncube
-
-cat > /etc/php/${php_version}/cli/conf.d/00-ioncube-loader.ini << EOF
-zend_extension=ioncube_loader_lin_${php_version}.so
+    cp ioncube_loader_lin_${php_version}.so /usr/lib/php/${php_ext_dir}
+    cd ..
+    rm -rf ioncube
+    cat > /etc/php/${php_version}/cli/conf.d/00-ioncube-loader.ini << EOF
+    zend_extension=ioncube_loader_lin_${php_version}.so
 EOF
+fi
+
+
 cd /opt/
 rm -rf DragonCore
 cd $HOME
